+++ src/common/POLFields/DetailsFields/FieldBinderyActive/index.js
@@ -1,0 +1,2 @@
+export { FieldBinderyActive } from './FieldBinderyActive';
+export { handleBinderyActiveFieldChange } from './handleBinderyActiveFieldChange';