+++ CHANGELOG.md
@@ -42,6 +42,7 @@
 * Include locations from member tenants in the CSV export file for orders. Refs UIOR-1467.
 * Display the total estimated price of PO in the actual currencies of the PO Lines, rather than in the converted value. Refs UIOR-1264.
 * Submit only fields registered in the form. Refs UIOR-1481.
+* Display a confirmation modal when activating "Bindery active" field in the PO Line form of post-pending order. Refs UIOR-1479.
 
 ## [8.0.5](https://github.com/folio-org/ui-orders/tree/v8.0.5) (2025-06-30)
 [Full Changelog](https://github.com/folio-org/ui-orders/compare/v8.0.4...v8.0.5)